LayoutTests/accessibility/gtk/xml-roles-exposed.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">
<div id="test">
 <div role="alert"></div>
 <div role="alertdialog"></div>
 <div role="application"></div>
 <div role="article"></div>
 <div role="banner"></div>
 <div role="blockquote"></div>
 <div role="button"></div>
 <div role="caption"></div>
 <div role="checkbox"></div>
 <div role="combobox"></div>
 <div role="complementary"></div>
 <div role="contentinfo"></div>
 <div role="definition"></div>
 <div role="dialog"></div>
 <div role="directory"></div>
 <div role="document"></div>
 <div role="feed"></div>
 <div role="figure"></div>
 <div role="form"></div>
 <div role="group"></div>
 <div role="heading"></div>
 <div role="img"></div>
 <div role="link"></div>
 <div role="log"></div>
 <div role="main"></div>
 <div role="marquee"></div>
 <div role="math"></div>
 <div role="navigation"></div>
 <div role="note"></div>
 <div role="paragraph"></div>
 <div role="presentation"></div>
 <div role="progressbar"></div>
 <div role="radio"></div>
 <div role="region" aria-label="X"></div>
 <div role="region">X</div>
 <div role="scrollbar"></div>
 <div role="search"></div>
 <div role="searchbox"></div>
 <div role="separator"></div>
 <div role="slider"></div>
 <div role="spinbutton"></div>
 <div role="status"></div>
 <div role="switch"></div>
 <div role="term"></div>
 <div role="textbox"></div>
 <div role="timer"></div>
 <div role="toolbar"></div>
 <div role="tooltip"></div>
 <div role="doc-abstract"></div>
 <div role="doc-acknowledgments"></div>
 <div role="doc-afterword"></div>
 <div role="doc-appendix"></div>
 <div role="doc-backlink"></div>
 <div role="doc-biblioentry"></div>
 <div role="doc-bibliography"></div>
 <div role="doc-biblioref"></div>
 <div role="doc-chapter"></div>
 <div role="doc-colophon"></div>
 <div role="doc-conclusion"></div>
 <div role="doc-cover"></div>
 <div role="doc-credit"></div>
 <div role="doc-credits"></div>
 <div role="doc-dedication"></div>
 <div role="doc-endnote"></div>
 <div role="doc-endnotes"></div>
 <div role="doc-epigraph"></div>
 <div role="doc-epilogue"></div>
 <div role="doc-errata"></div>
 <div role="doc-example"></div>
 <div role="doc-footnote"></div>
 <div role="doc-foreword"></div>
 <div role="doc-glossary"></div>
 <div role="doc-glossref"></div>
 <div role="doc-index"></div>
 <div role="doc-introduction"></div>
 <div role="doc-noteref"></div>
 <div role="doc-notice"></div>
 <div role="doc-pagebreak"></div>
 <div role="doc-pagelist"></div>
 <div role="doc-part"></div>
 <div role="doc-preface"></div>
 <div role="doc-prologue"></div>
 <div role="doc-pullquote"></div>
 <div role="doc-qna"></div>
 <div role="doc-subtitle"></div>
 <div role="doc-tip"></div>
 <div role="doc-toc"></div>
 <div role="graphics-document"></div>
 <div role="graphics-object"></div>
 <div role="graphics-symbol"></div>
</div>
<p id="description"></p>
<div id="console"></div>
<script>
description("Verify exposure of ARIA role values via the 'xml-roles' object attribute.");
if (window.accessibilityController) {
    document.getElementById("body").focus();
    debug(accessibilityController.focusedElement.attributesOfChildren());
    document.getElementById("test").style.visibility = "hidden";
}
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
